source/repos/LandingPage/KonnectLandingPage/requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.9
django-admin-interface==0.18.5
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-embed-video==1.4.1
django-etc==1.3.1
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-hitcount==1.3.5
django-js-asset==1.2.2
django-next-prev==1.1.0
django-richtextfield==1.6.1
django-taggit==2.1.0
django-user-messages==0.7.0
graphene==2.1.9
graphene-django==2.14.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
Pillow==9.0.0
promise==2.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.8
